pyproject.toml
[project]
name = "storyswarm-backend"
version = "0.1.0"
description = "Backend FastAPI application for storyswarm"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pydantic>=2.11.2",
    "pydantic-ai>=0.1.3",
    "structlog>=25.2.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "python-dotenv>=1.1.0",
    "types-pyyaml>=6.0.12.20250402",
]

[tool.mypy]
plugins = ['pydantic.mypy']
